move FirstBoot into overlord/boot
