+++ overlord/export_test.go
@@ -53,5 +53,3 @@ func MockEnsureNext(o *Overlord, t time.Time) {
 func (o *Overlord) Engine() *StateEngine {
 	return o.stateEng
 }
-
-var PopulateStateFromInstalled = populateStateFromInstalled
